pyproject.toml
[tool.poetry]
name = "automation-framework"
version = "0.1.0"
description = "Enterprise API Testing Framework"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.0"
requests = "^2.31.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-xdist = "^3.3.0"
pytest-env = "^1.0.0"
pytest-html = "^4.1.0"
allure-pytest = "^2.13.0"
pymysql = "^1.1.0"
pymongo = "^4.5.0"
redis = "^5.0.0"
python-memcached = "^1.59"
pyyaml = "^6.0.0"
python-dotenv = "^1.0.0"
jsonschema = "^4.19.0"
faker = "^19.3.0"
arrow = "^1.3.0"
structlog = "^23.1.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
prometheus-client = "^0.17.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --asyncio-mode=auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "mysql: marks tests that require MySQL",
    "mongodb: marks tests that require MongoDB",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 